Moved channel tests from api app to channels app
